UI enhancements

+++ CarpoolingCR/Models/UserRating.cs
@@ -12,7 +12,7 @@ namespace CarpoolingCR.Models
         public int TripId { get; set; }
         public string FromId { get; set; }
         public string ToId { get; set; }
-        public int Stars { get; set; }
+        public decimal Stars { get; set; }
         public string Comments { get; set; }
         public DateTime DateTime { get; set; }
